requirements.txt
boto3
botocore
pandas
psycopg2_binary
python-dotenv
python_dateutil
Requests
